EasyProg: Cleaned up startup code
Some cleanup of the startup code. Additionally we have
a small wait loop at the beginning of the reset vector.
This should make sure everything is up and running in
the c64 in case the cpu starts executing before all chips
leave the reset state. Additionally this makes sure the
RAM chips are fully working. The RAM data sheets say that
the RAM must be powered on 100 or 200 us and that there
must be at least 8 RAS cycles before it is fully working.